Move TravisLintBear.py to configfiles directory
Closes https://github.com/coala/coala-bears/issues/1979
