+++ bears/configfiles/TravisLintBear.py
